# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ParkSecure API",
    "version": "1.0.0",
    "description": "REST API for the ParkSecure vehicle security platform.\n\n**Authentication:** obtain a JWT via `POST /api/auth/login` and send it as `Authorization: Bearer <token>`.\n\n**Roles:** `super_admin`, `park_admin`, `facility_admin`, `security`, `vehicle_owner`. Staff with an assigned facility are automatically scoped to that facility; vehicle owners only see their own records."
  },
  "servers": [
    {
      "url": "https://parksecure-api.onrender.com"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Facility": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "billing_mode": {
            "type": "string",
            "enum": [
              "free",
              "standard",
              "hourly"
            ]
          },
          "hourly_rate": {
            "type": "number"
          },
          "standard_rate": {
            "type": "number"
          },
          "subscription_monthly_fee": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "exit_timeout_minutes": {
            "type": "number"
          },
          "qr_deactivate_after_days": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "logo_url": {
            "type": "string"
          },
          "contact_phone": {
            "type": "string"
          },
          "contact_email": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "qr_code_id": {
            "type": "string"
          },
          "owner_name": {
            "type": "string"
          },
          "owner_phone": {
            "type": "string"
          },
          "owner_email": {
            "type": "string"
          },
          "owner_photo_url": {
            "type": "string"
          },
          "driver_name": {
            "type": "string"
          },
          "driver_phone": {
            "type": "string"
          },
          "plate_number": {
            "type": "string"
          },
          "make_model": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "photo_front_url": {
            "type": "string"
          },
          "photo_back_url": {
            "type": "string"
          },
          "photo_side_url": {
            "type": "string"
          },
          "photo_url": {
            "type": "string"
          },
          "facility_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "blacklisted",
              "suspended"
            ]
          },
          "registration_type": {
            "type": "string",
            "enum": [
              "permanent",
              "guest"
            ]
          },
          "guest_pass_expires": {
            "type": "string",
            "format": "date-time"
          },
          "guest_pass_notified": {
            "type": "boolean"
          },
          "is_inside": {
            "type": "boolean"
          },
          "last_entry": {
            "type": "string",
            "format": "date-time"
          },
          "last_exit": {
            "type": "string",
            "format": "date-time"
          },
          "registered": {
            "type": "boolean"
          },
          "qr_requested": {
            "type": "boolean"
          },
          "qr_request_date": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "plate_number"
        ]
      },
      "QRCode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "code_id": {
            "type": "string"
          },
          "facility_id": {
            "type": "string"
          },
          "batch_name": {
            "type": "string"
          },
          "code_type": {
            "type": "string",
            "enum": [
              "permanent",
              "guest"
            ]
          },
          "guest_duration_hours": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "assigned",
              "deactivated"
            ]
          },
          "vehicle_id": {
            "type": "string"
          },
          "generated_by": {
            "type": "string"
          },
          "qr_image_url": {
            "type": "string"
          }
        },
        "required": [
          "code_id",
          "facility_id"
        ]
      },
      "ScanLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "vehicle_id": {
            "type": "string"
          },
          "qr_code_id": {
            "type": "string"
          },
          "facility_id": {
            "type": "string"
          },
          "scan_type": {
            "type": "string",
            "enum": [
              "entry",
              "exit"
            ]
          },
          "scanned_by": {
            "type": "string"
          },
          "scanned_by_name": {
            "type": "string"
          },
          "plate_number": {
            "type": "string"
          },
          "owner_name": {
            "type": "string"
          },
          "driver_name": {
            "type": "string"
          },
          "driver_phone": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "pending_approval",
              "denied",
              "timeout"
            ]
          },
          "items_declared": {
            "type": "array",
            "items": {}
          },
          "booth_photo_urls": {
            "type": "array",
            "items": {}
          },
          "security_confirmed": {
            "type": "boolean"
          },
          "security_notes": {
            "type": "string"
          },
          "discrepancies": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "qr_code_id",
          "scan_type"
        ]
      },
...
      "ExitRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "vehicle_id": {
            "type": "string"
          },
          "qr_code_id": {
            "type": "string"
          },
          "facility_id": {
            "type": "string"
          },
          "owner_name": {
            "type": "string"
          },
          "owner_phone": {
            "type": "string"
          },
          "owner_email": {
            "type": "string"
          },
          "plate_number": {
            "type": "string"
          },
          "requested_by": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "timeout",
              "escalated"
            ]
          },
          "responded_at": {
            "type": "string",
            "format": "date-time"
          },
          "responded_by": {
            "type": "string"
          },
          "approval_token": {
            "type": "string"
          },
          "billing_amount": {
            "type": "number"
          },
          "duration_minutes": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "vehicle_id",
          "facility_id"
        ]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/scans/lookup": {
      "get": {
        "tags": [
          "Scanning"
        ],
        "summary": "Look up a scanned QR code",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "PSK-0001"
          }
        ],
        "responses": {
          "200": {
            "description": "Lookup result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "ok",
                        "unknown_code",
                        "unregistered",
                        "blacklisted",
                        "guest_expired"
                      ]
                    },
                    "qr_code": {
                      "$ref": "#/components/schemas/QRCode"
                    },
                    "vehicle": {
                      "$ref": "#/components/schemas/Vehicle"
                    },
                    "facility": {
                      "$ref": "#/components/schemas/Facility"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/scans/entry": {
      "post": {
        "tags": [
          "Scanning"
        ],
        "summary": "Record a vehicle entry",
        "description": "Creates the entry ScanLog + ItemLogs and marks the vehicle inside. Denies blacklisted vehicles and expired guest passes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "qr_code_id": {
                    "type": "string"
                  },
                  "vehicle_id": {
                    "type": "string"
                  },
                  "driver_name": {
                    "type": "string"
                  },
                  "driver_phone": {
                    "type": "string"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "number"
                        },
                        "description": {
                          "type": "string"
                        },
                        "photo_urls": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "booth_photo_urls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entry recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scan_log": {
                      "$ref": "#/components/schemas/ScanLog"
                    },
                    "vehicle": {
                      "$ref": "#/components/schemas/Vehicle"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Blacklisted or expired guest pass",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Vehicle already inside",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/exit": {
      "post": {
        "tags": [
          "Scanning"
        ],
        "summary": "Record a vehicle exit (starts the approval workflow)",
        "description": "Computes the parking fee from the facility billing mode, creates an ExitRequest with an approval token and notifies the owner (push + email). Auto-approves when the owner enabled security override.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "qr_code_id": {
                    "type": "string"
                  },
                  "vehicle_id": {
                    "type": "string"
                  },
                  "items_state": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "number"
                        },
                        "confirmed": {
                          "type": "boolean"
                        },
                        "discrepancy": {
                          "type": "string"
                        },
                        "photo_urls": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "extra_items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "number"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "booth_photo_urls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exit recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exit_request": {
                      "$ref": "#/components/schemas/ExitRequest"
                    },
                    "scan_log": {
                      "$ref": "#/components/schemas/ScanLog"
                    },
                    "auto_approved": {
                      "type": "boolean"
                    },
                    "billing": {
                      "type": "object",
                      "properties": {
                        "amount": {
                          "type": "number"
                        },
                        "durationMinutes": {
                          "type": "number"
                        },
                        "mode": {
                          "type": "string"
                        },
                        "currency": {
                          "type": "string"
                        }
                      }
                    },
                    "whatsapp_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "discrepancies": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}